moved artifact types
