+++ src/golang/lib/collections/artifact/boolean/bool.go
