+++ src/golang/lib/collections/artifact/float/float.go
