+++ src/golang/lib/collections/tests/server_custom_test.go
@@ -2,13 +2,13 @@ package tests
 
 import (
 	"context"
+	"github.com/aqueducthq/aqueduct/lib/collections/artifact/table"
 	"testing"
 
 	"github.com/aqueducthq/aqueduct/cmd/server/queries"
 	"github.com/aqueducthq/aqueduct/lib/collections/artifact"
 	"github.com/aqueducthq/aqueduct/lib/collections/integration"
 	"github.com/aqueducthq/aqueduct/lib/collections/operator"
-	"github.com/aqueducthq/aqueduct/lib/workflow/artifact/table"
 	"github.com/aqueducthq/aqueduct/lib/workflow/operator/connector"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
